scripts/automation/automerge/requirements.txt
pyyaml
PyGithub
GitPython